requirements.txt
PySide6>=6.6
psycopg2-binary>=2.9
scikit-learn>=1.2.2
numpy>=1.23
scipy>=1.10
PyPDF2>=3.0.0
pdfplumber>=0.9.0 